Move to IT, add localstack to docker-compose
