+++ docker-compose.yml
@@ -54,11 +54,26 @@ services:
       - "5432:5432"
     volumes:
       - db_data:/var/lib/postgresql/data
+  # localstack is used for testing
+  localstack:
+    image: localstack/localstack:1.3.1
+    container_name: localstack1
+    ports:
+      - "127.0.0.1:4566:4566"            # LocalStack Gateway
+      - "127.0.0.1:4510-4559:4510-4559"  # external services port range
+    environment:
+      - DEBUG=${DEBUG-}
+      - LAMBDA_EXECUTOR=${LAMBDA_EXECUTOR-}
+      - DOCKER_HOST=unix:///var/run/docker.sock
+    volumes:
+      - localstack_data:/var/lib/localstack
+      - "/var/run/docker.sock:/var/run/docker.sock"
 
 volumes:
   data01:
     driver: local
   db_data:
+  localstack_data:
 
 networks:
   elastic:
