+++ dockstore-integration-testing/src/test/resources/prototype-metrics-request-body.json
